Fix #PAMP-3 Stage Review Resolve issue
